Refactor location and distance tracking logic
Replaced GMS/HMS location dependencies with Amap-only implementation and updated the AndroidManifest API key. Refactored LocationProvider to use a custom KalmanFilter for smoother GPS data, changed totalDistance type from Double to Float, and unified distance tracking logic. Updated related usages in InspectionManager and DashboardUpdater. Added a new KalmanFilter utility class for GPS smoothing. Restored AddressProvider implementation.

+++ app/src/main/java/com/example/roadinspection/data/model/HighFrequencyData.kt
@@ -13,5 +13,5 @@ data class HighFrequencyData(
     val timeDiff: Long = 0L,
     val lat: Double = 0.0,
     val lng: Double = 0.0,
-    val totalDistance: Double = 0.0
+    val totalDistance: Float = 0f
 )
